+++ safer_pay/doc/RELEASE_NOTES.md
@@ -4,4 +4,4 @@
 #### Version 15.0.1.0.0
 #### ADD
 
-- Initial commit for Safer-pay Payment Gateway Integration
+- Initial commit for Safer-Pay Payment Gateway Integration